server-requirements.txt
bcrypt==3.2.0
cryptography==3.4.7
fastapi==0.63.0
passlib==1.7.4
pymongo==3.11.3
pysodium==0.7.9
python-jose==3.2.0
python-multipart==0.0.5



